# web/frontend/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelAvailabilityRequest": {
        "additionalProperties": false,
        "properties": {
          "api_key": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Api Key",
            "type": "string"
          },
          "base_url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Base Url",
            "type": "string"
          },
          "default_body": {
            "additionalProperties": true,
            "title": "Default Body",
            "type": "object"
          },
          "model_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Model Name",
            "type": "string"
          },
          "protocol": {
            "$ref": "#/components/schemas/ModelProviderProtocol",
            "default": "OPENAI_COMPATIBLE"
          },
          "proxy_mode": {
            "$ref": "#/components/schemas/ModelProviderProxyMode",
            "default": "SYSTEM"
          },
          "proxy_password": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Password"
          },
          "proxy_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Url"
          },
          "proxy_username": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Username"
          },
          "verify_ssl": {
            "default": true,
            "title": "Verify Ssl",
            "type": "boolean"
          }
        },
        "required": [
          "base_url",
          "api_key",
          "model_name"
        ],
        "title": "ModelAvailabilityRequest",
        "type": "object"
      },
      "ModelAvailabilityResponse": {
        "additionalProperties": false,
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "output": {
            "title": "Output"
          },
          "response_body": {
            "title": "Response Body",
            "type": "string"
          },
          "status_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Code"
          }
        },
        "required": [
          "available",
          "latency_ms",
          "status_code",
          "response_body",
          "error"
        ],
        "title": "ModelAvailabilityResponse",
        "type": "object"
      },
      "ModelProviderConfiguration": {
        "additionalProperties": false,
        "properties": {
          "api_key": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Api Key",
            "type": "string"
          },
          "base_url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Base Url",
            "type": "string"
          },
          "model_configs": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ModelRuntimeConfiguration"
            },
            "title": "Model Configs",
            "type": "object"
          },
          "model_endpoint": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Endpoint"
          },
          "models": {
            "items": {
              "type": "string"
            },
            "maxItems": 500,
            "title": "Models",
            "type": "array"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "protocol": {
            "$ref": "#/components/schemas/ModelProviderProtocol"
          },
          "proxy_mode": {
            "$ref": "#/components/schemas/ModelProviderProxyMode",
            "default": "SYSTEM"
          },
          "proxy_password": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Password"
          },
          "proxy_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Url"
          },
          "proxy_username": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Username"
          },
          "verify_ssl": {
            "default": true,
            "title": "Verify Ssl",
            "type": "boolean"
          },
          "website_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website Url"
          }
        },
        "required": [
          "api_key",
          "base_url",
          "protocol",
          "models"
        ],
        "title": "ModelProviderConfiguration",
        "type": "object"
      },
      "ModelProviderProtocol": {
        "enum": [
          "OPENAI_COMPATIBLE",
          "OPENAI_RESPONSES",
          "ANTHROPIC",
          "MANUAL"
        ],
        "title": "ModelProviderProtocol",
        "type": "string"
      },
...
      "ProviderConnectionRequest": {
        "additionalProperties": false,
        "properties": {
          "api_key": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Api Key",
            "type": "string"
          },
          "base_url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Base Url",
            "type": "string"
          },
          "protocol": {
            "$ref": "#/components/schemas/ModelProviderProtocol",
            "default": "OPENAI_COMPATIBLE"
          },
          "proxy_mode": {
            "$ref": "#/components/schemas/ModelProviderProxyMode",
            "default": "SYSTEM"
          },
          "proxy_password": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Password"
          },
          "proxy_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Url"
          },
          "proxy_username": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Username"
          },
          "verify_ssl": {
            "default": true,
            "title": "Verify Ssl",
            "type": "boolean"
          }
        },
        "required": [
          "base_url",
          "api_key"
        ],
        "title": "ProviderConnectionRequest",
        "type": "object"
      },
...
      "ProviderEnvelope": {
        "additionalProperties": false,
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/ModelProviderRecord"
          }
        },
        "required": [
          "provider"
        ],
        "title": "ProviderEnvelope",
        "type": "object"
      },
      "ProviderLatencyResponse": {
        "additionalProperties": false,
        "properties": {
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "reachable": {
            "title": "Reachable",
            "type": "boolean"
          },
          "status_code": {
            "title": "Status Code",
            "type": "integer"
          }
        },
        "required": [
          "reachable",
          "latency_ms",
          "status_code"
        ],
        "title": "ProviderLatencyResponse",
        "type": "object"
      },
      "ProviderListResponse": {
        "additionalProperties": false,
        "properties": {
          "providers": {
            "items": {
              "$ref": "#/components/schemas/ModelProviderSummary"
            },
            "title": "Providers",
            "type": "array"
          }
        },
        "required": [
          "providers"
        ],
        "title": "ProviderListResponse",
        "type": "object"
      },
      "ProviderModelsResponse": {
        "additionalProperties": false,
        "properties": {
          "endpoint": {
            "title": "Endpoint",
            "type": "string"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ProviderDiscoveredModel"
            },
            "title": "Models",
            "type": "array"
          },
          "protocol": {
            "title": "Protocol",
            "type": "string"
          }
        },
        "required": [
          "protocol",
          "endpoint",
          "latency_ms",
          "models"
        ],
        "title": "ProviderModelsResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "EvalFlow",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/model-providers": {
      "get": {
        "operationId": "list_providers_api_model_providers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Providers",
        "tags": [
          "model-providers"
        ]
      },
      "post": {
        "operationId": "create_provider_api_model_providers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelProviderConfiguration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Provider",
        "tags": [
          "model-providers"
        ]
      }
    },
    "/api/model-providers/latency": {
      "post": {
        "operationId": "test_latency_api_model_providers_latency_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderLatencyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Latency",
        "tags": [
          "model-providers"
        ]
      }
    },
    "/api/model-providers/models": {
      "post": {
        "operationId": "fetch_models_api_model_providers_models_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderModelsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch Models",
        "tags": [
          "model-providers"
        ]
      }
    },
    "/api/model-providers/test-model": {
      "post": {
        "operationId": "test_model_availability_api_model_providers_test_model_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelAvailabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelAvailabilityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Model Availability",
        "tags": [
          "model-providers"
        ]
      }
    },
    "/api/model-providers/{provider_id}": {
      "delete": {
        "operationId": "delete_provider_api_model_providers__provider_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderSummaryEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Provider",
        "tags": [
          "model-providers"
        ]
      },
      "get": {
        "operationId": "get_provider_api_model_providers__provider_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Provider",
        "tags": [
          "model-providers"
        ]
      },
      "put": {
        "operationId": "update_provider_api_model_providers__provider_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelProviderConfiguration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Provider",
        "tags": [
          "model-providers"
        ]
      }
    },
...
  }
}